# boot/assets/sw/config/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is builtin RK common service.",
        "title": "RK Common Service",
        "contact": {
            "name": "rk-dev",
            "url": "https://github.com/rookie-ninja/rk-echo",
            "email": "lark@pointgoal.io"
        },
        "license": {
            "name": "Apache 2.0 License",
            "url": "https://github.com/rookie-ninja/rk-echo/blob/master/LICENSE.txt"
        },
        "version": "1.0"
    },
    "paths": {
        "/rk/v1/apis": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List API",
                "operationId": "5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.ApisResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/certs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List CertEntry",
                "operationId": "9",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.CertsResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/configs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List ConfigEntry",
                "operationId": "4",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.ConfigsResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/deps": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List dependencies related application",
                "operationId": "11",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.DepResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/entries": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List all Entry",
                "operationId": "8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.EntriesResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/gc": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Trigger Gc",
                "operationId": "2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.GcResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/git": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Git information.",
                "operationId": "14",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.GitResponse"
                        }
                    }
                }
            }
        },
...
        "/rk/v1/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get application and process info",
                "operationId": "3",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.ProcessInfo"
                        }
                    }
                }
            }
        },
        "/rk/v1/license": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get license related application",
                "operationId": "12",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.LicenseResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/logs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List logger related entries",
                "operationId": "10",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.LogsResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/readme": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get README file.",
                "operationId": "13",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.ReadmeResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/req": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List prometheus metrics of requests",
                "operationId": "7",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.ReqResponse"
                        }
                    }
                }
            }
        },
        "/rk/v1/sys": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWT": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get OS Stat",
                "operationId": "6",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rkentry.SysResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "rkentry.ApisResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rkentry.ApisResponse_Entry"
                    }
                }
            }
        },
...
        "rkentry.CertsResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rkentry.CertsResponse_Entry"
                    }
                }
            }
        },
...
        "rkentry.ConfigsResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rkentry.ConfigsResponse_ConfigEntry"
                    }
                }
            }
        },
...
        "rkentry.DepResponse": {
            "type": "object",
            "properties": {
                "goMod": {
                    "type": "string"
                }
            }
        },
        "rkentry.EntriesResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/rkentry.EntriesResponse_Entry"
                        }
                    }
                }
            }
        },
...
        "rkentry.LicenseResponse": {
            "type": "object",
            "properties": {
                "license": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}